feat: evt based window (#138)
## Purpose of Changes and their Description

Rework submission logic by subscribing to rpc web socket
`EventReputerSubmissionWindowOpened` &
`EventWorkerSubmissionWindowOpened` events, instead of calculating the
time the window will open.

In case the web socket connection is lost the underlying cometbft client
will try to reconnect until some max attempt, which once reached will
stop the whole app.

All the logic related to the triggering and websocket subscription are
abstracted and contained in the `WindowWaker` type.

A `SubmissionWindowManager` allows to create the wakers and to properly
manage stopping them upon shutdown.

This has been tested locally.

## Link(s) to Ticket(s) or Issue(s) resolved by this PR

## Are these changes tested and documented?

- [x] If tested, please describe how. If not, why tests are not needed.
- [x] If documented, please describe where. If not, describe why docs
are not needed.
- [x] Added to `Unreleased` section of `CHANGELOG.md`?